# web: integrate `DllPlugin` in Storybook development build