--- a/client/storybook/package.json
+++ b/client/storybook/package.json
@@ -11,7 +11,7 @@
     "build": "TS_NODE_TRANSPILE_ONLY=true build-storybook -c ./src -s ./assets",
     "build:webpack-stats": "TS_NODE_TRANSPILE_ONLY=true WEBPACK_DLL_PLUGIN=false start-storybook -c ./src -s ./assets --smoke-test --webpack-stats-json ./storybook-static --loglevel warn",
     "build:dll-bundle": "TS_NODE_TRANSPILE_ONLY=true webpack --config ./src/webpack.config.dll.ts --no-stats",
-    "start:dll": "TS_NODE_TRANSPILE_ONLY=true WEBPACK_DLL_PLUGIN=true start-storybook -p 9001 -c ./src -s ./assets --loglevel warn",
+    "start:dll": "TS_NODE_TRANSPILE_ONLY=true WEBPACK_DLL_PLUGIN=true start-storybook -p 9001 -c ./src -s ./assets",
     "clean:dll": "rm -rf assets/dll-bundle storybook-static/*-stats.json",
     "test": "jest"
   }