--- a/client/storybook/src/dllPlugin/index.ts
+++ b/client/storybook/src/dllPlugin/index.ts
@@ -1,0 +1,3 @@
+export * from './ensureDllBundleIsReady'
+export * from './getVendorModules'
+export * from './getWebpackStats'