--- a/client/storybook/src/environment-config.ts
+++ b/client/storybook/src/environment-config.ts
@@ -2,6 +2,8 @@ const getEnvironmentBoolean = (value = 'false'): boolean => Boolean(JSON.parse(v
 
 export const environment = {
     customStoriesGlob: process.env.STORIES_GLOB,
+    isDLLPluginEnabled: getEnvironmentBoolean(process.env.WEBPACK_DLL_PLUGIN),
+    isProgressPluginEnabled: getEnvironmentBoolean(process.env.WEBPACK_PROGRESS_PLUGIN),
     isBundleAnalyzerEnabled: getEnvironmentBoolean(process.env.WEBPACK_BUNDLE_ANALYZER),
     isSpeedAnalyzerEnabled: getEnvironmentBoolean(process.env.WEBPACK_SPEED_ANALYZER),
     shouldMinify: getEnvironmentBoolean(process.env.MINIFY),