--- a/client/storybook/src/main.ts
+++ b/client/storybook/src/main.ts
@@ -1,16 +1,28 @@
 import path from 'path'
 
+import { Options } from '@storybook/core-common'
+import CaseSensitivePathsPlugin from 'case-sensitive-paths-webpack-plugin'
 import { remove } from 'lodash'
-import MonacoWebpackPlugin from 'monaco-editor-webpack-plugin'
+import signale from 'signale'
 import SpeedMeasurePlugin from 'speed-measure-webpack-plugin'
 import TerserPlugin from 'terser-webpack-plugin'
-import { Configuration, DefinePlugin, ProgressPlugin, RuleSetUseItem, RuleSetUse } from 'webpack'
+import { DllReferencePlugin, Configuration, DefinePlugin, ProgressPlugin, RuleSetUseItem, RuleSetUse } from 'webpack'
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer'
 
+import { ensureDllBundleIsReady } from './dllPlugin'
 import { environment } from './environment-config'
-
-const rootPath = path.resolve(__dirname, '../../../')
-const monacoEditorPaths = [path.resolve(rootPath, 'node_modules', 'monaco-editor')]
+import {
+    rootPath,
+    monacoEditorPath,
+    dllPluginConfig,
+    dllBundleManifestPath,
+    getMonacoCSSRule,
+    getMonacoTTFRule,
+    getMonacoWebpackPlugin,
+    nodeModulesPath,
+    getBasicCSSLoader,
+    readJsonFile,
+} from './webpack.config.common'
 
 const getStoriesGlob = (): string[] => {
     if (process.env.STORIES_GLOB) {
@@ -44,6 +56,18 @@ const getCSSLoaders = (...loaders: RuleSetUseItem[]): RuleSetUse => [
     },
 ]
 
+const getDllScriptTag = (): string => {
+    ensureDllBundleIsReady()
+    signale.await('Waiting for Webpack to compile Storybook preview.')
+
+    const dllManifest = readJsonFile(dllBundleManifestPath) as Record<string, string>
+
+    return `
+        <!-- Load JS bundle created by DLL_PLUGIN  -->
+        <script src="/dll-bundle/${dllManifest['dll.js']}"></script>
+    `
+}
+
 const config = {
     stories: getStoriesGlob(),
     addons: [
@@ -66,19 +90,23 @@ const config = {
         reactDocgen: false,
     },
 
-    webpackFinal: (config: Configuration) => {
+    // Include DLL bundle script tag into preview-head.html if DLLPlugin is enabled.
+    previewHead: (head: string) => `
+        ${head}
+        ${environment.isDLLPluginEnabled ? getDllScriptTag() : ''}
+    `,
+
+    webpackFinal: (config: Configuration, options: Options) => {
+        config.stats = 'errors-warnings'
         config.mode = environment.shouldMinify ? 'production' : 'development'
 
         // Check the default config is in an expected shape.
-        if (!config.module) {
+        if (!config.module || !config.plugins) {
             throw new Error(
                 'The format of the default storybook webpack config changed, please check if the config in ./src/main.ts is still valid'
             )
         }
 
-        if (!config.plugins) {
-            config.plugins = []
-        }
         config.plugins.push(
             new DefinePlugin({
                 NODE_ENV: JSON.stringify(config.mode),
@@ -106,10 +134,6 @@ const config = {
             ]
         }
 
-        if (process.env.CI) {
-            remove(config.plugins, plugin => plugin instanceof ProgressPlugin)
-        }
-
         // We don't use Storybook's default Babel config for our repo, it doesn't include everything we need.
         config.module.rules.splice(0, 1)
         config.module.rules.unshift({
@@ -121,40 +145,21 @@ const config = {
             },
         })
 
-        config.plugins.push(
-            new MonacoWebpackPlugin({
-                languages: ['json'],
-                features: [
-                    'bracketMatching',
-                    'clipboard',
-                    'coreCommands',
-                    'cursorUndo',
-                    'find',
-                    'format',
-                    'hover',
-                    'inPlaceReplace',
-                    'iPadShowKeyboard',
-                    'links',
-                    'suggest',
-                ],
-            })
-        )
-
-        const storybookDirectory = path.resolve(rootPath, 'node_modules/@storybook')
+        const storybookPath = path.resolve(nodeModulesPath, '@storybook')
 
         // Put our style rules at the beginning so they're processed by the time it
         // gets to storybook's style rules.
         config.module.rules.unshift({
             test: /\.(sass|scss)$/,
             // Make sure Storybook styles get handled by the Storybook config
-            exclude: [/\.module\.(sass|scss)$/, storybookDirectory],
-            use: getCSSLoaders('@terminus-term/to-string-loader', { loader: 'css-loader', options: { url: false } }),
+            exclude: [/\.module\.(sass|scss)$/, storybookPath],
+            use: getCSSLoaders('@terminus-term/to-string-loader', getBasicCSSLoader()),
         })
 
         config.module?.rules.unshift({
             test: /\.(sass|scss)$/,
             include: /\.module\.(sass|scss)$/,
-            exclude: storybookDirectory,
+            exclude: storybookPath,
             use: getCSSLoaders('style-loader', {
                 loader: 'css-loader',
                 options: {
@@ -173,30 +178,14 @@ const config = {
         if (!cssRule) {
             throw new Error('Cannot find original CSS rule')
         }
-        cssRule.include = storybookDirectory
+        cssRule.include = storybookPath
 
         config.module.rules.push({
             // CSS rule for external plain CSS (skip SASS and PostCSS for build perf)
             test: /\.css$/,
             // Make sure Storybook styles get handled by the Storybook config
-            exclude: [storybookDirectory, ...monacoEditorPaths],
-            use: ['@terminus-term/to-string-loader', { loader: 'css-loader', options: { url: false } }],
-        })
-
-        config.module.rules.push({
-            // CSS rule for monaco-editor, it expects styles to be loaded with `style-loader`.
-            test: /\.css$/,
-            include: monacoEditorPaths,
-            // Make sure Storybook styles get handled by the Storybook config
-            exclude: [storybookDirectory],
-            use: ['style-loader', { loader: 'css-loader' }],
-        })
-
-        config.module?.rules.unshift({
-            // TTF rule for monaco-editor
-            test: /\.ttf$/,
-            include: monacoEditorPaths,
-            use: ['file-loader'],
+            exclude: [storybookPath, monacoEditorPath],
+            use: ['@terminus-term/to-string-loader', getBasicCSSLoader()],
         })
 
         config.module.rules.push({
@@ -204,10 +193,32 @@ const config = {
             use: ['raw-loader'],
         })
 
-        Object.assign(config.entry, {
-            'editor.worker': 'monaco-editor/esm/vs/editor/editor.worker.js',
-            'json.worker': 'monaco-editor/esm/vs/language/json/json.worker',
-        })
+        // Disable `CaseSensitivePathsPlugin` by default to speed up development build.
+        // Similar discussion: https://github.com/vercel/next.js/issues/6927#issuecomment-480579191
+        remove(config.plugins, plugin => plugin instanceof CaseSensitivePathsPlugin)
+
+        // Disable `ProgressPlugin` by default to speed up development build.
+        // Can be re-enabled by setting `WEBPACK_PROGRESS_PLUGIN` env variable.
+        if (!environment.isProgressPluginEnabled) {
+            remove(config.plugins, plugin => plugin instanceof ProgressPlugin)
+        }
+
+        if (environment.isDLLPluginEnabled && !options.webpackStatsJson) {
+            config.plugins.unshift(
+                new DllReferencePlugin({
+                    context: dllPluginConfig.context,
+                    manifest: dllPluginConfig.path,
+                })
+            )
+        } else {
+            config.plugins.push(getMonacoWebpackPlugin())
+            config.module.rules.push(getMonacoCSSRule(), getMonacoTTFRule())
+
+            Object.assign(config.entry, {
+                'editor.worker': 'monaco-editor/esm/vs/editor/editor.worker.js',
+                'json.worker': 'monaco-editor/esm/vs/language/json/json.worker',
+            })
+        }
 
         if (environment.isBundleAnalyzerEnabled) {
             config.plugins.push(new BundleAnalyzerPlugin())