--- a/client/storybook/src/webpack.config.dll.ts
+++ b/client/storybook/src/webpack.config.dll.ts
@@ -1,0 +1,48 @@
+import signale from 'signale'
+import { DllPlugin, Configuration } from 'webpack'
+import { WebpackManifestPlugin } from 'webpack-manifest-plugin'
+
+import { getWebpackStats, getVendorModules } from './dllPlugin'
+import {
+    monacoEditorPath,
+    dllPluginConfig,
+    getMonacoCSSRule,
+    getMonacoTTFRule,
+    getMonacoWebpackPlugin,
+    getBasicCSSLoader,
+    dllBundleManifestPath,
+} from './webpack.config.common'
+
+const webpackStats = getWebpackStats()
+signale.await('Waiting for Webpack to build DLL bundle based on vendor stats.')
+
+const config: Configuration = {
+    mode: 'development',
+    stats: 'errors-warnings',
+    entry: {
+        dll: [...getVendorModules(webpackStats), 'monaco-editor'],
+    },
+    output: {
+        filename: '[name].bundle.[contenthash].js',
+        path: dllPluginConfig.context,
+        library: dllPluginConfig.name,
+    },
+    module: {
+        rules: [
+            getMonacoCSSRule(),
+            getMonacoTTFRule(),
+            {
+                test: /\.css$/,
+                exclude: [monacoEditorPath],
+                use: [getBasicCSSLoader()],
+            },
+        ],
+    },
+    plugins: [
+        getMonacoWebpackPlugin(),
+        new DllPlugin(dllPluginConfig),
+        new WebpackManifestPlugin({ fileName: dllBundleManifestPath }),
+    ],
+}
+
+module.exports = config