--- a/package.json
+++ b/package.json
@@ -136,6 +136,7 @@
     "@testing-library/react-hooks": "^3.4.1",
     "@types/babel__core": "7.1.12",
     "@types/bloomfilter": "^0.0.0",
+    "@types/case-sensitive-paths-webpack-plugin": "^2.1.5",
     "@types/chai": "4.2.14",
     "@types/chai-as-promised": "7.1.3",
     "@types/chrome": "0.0.127",